+++ modules/basegl/processors/imageprocessing/imagelayoutgl.cpp
@@ -209,16 +209,16 @@ void ImageLayoutGL::updateViewports(ivec2 dim, bool force) {
     if (!force && (currentDim_ == dim) && (currentLayout_ == layout_.get())) return;  // no changes
 
     viewManager_.clear();
-    const int smallWindowHeight = dim.y / 3.0f;
+    const int smallWindowHeight = static_cast<int>(dim.y / 3.0f);
 
     const int extra1 = dim.y % 3 >= 1 ? 1 : 0;  // add extra pixels to the small "windows" if the
     const int extra2 = dim.y % 3 >= 2 ? 1 : 0;  // size is not divisible by 3 to avoid black borders
 
-    const int midx = verticalSplitter_ * dim.x;
-    const int midy = horizontalSplitter_ * dim.y;
+    const int midx = static_cast<int>(verticalSplitter_ * dim.x);
+    const int midy = static_cast<int>(horizontalSplitter_ * dim.y);
 
-    const int leftWindow3L1RX = vertical3Left1RightSplitter_ * dim.x;
-    const int leftWindow3R1LX = vertical3Right1LeftSplitter_ * dim.x;
+    const int leftWindow3L1RX = static_cast<int>(vertical3Left1RightSplitter_ * dim.x);
+    const int leftWindow3R1LX = static_cast<int>(vertical3Right1LeftSplitter_ * dim.x);
 
     const int portCount = static_cast<int>(multiinport_.getConnectedOutports().size());
